Instead of setting the geometry on the widget set it to the platform window
When setting the geometry on the QGLWidget we have to call show() to make sure screen size
is taken over by the widget. If we're setting the geomtry instead on the QPlatformWindow
the geometry is set as before to the widget but for the native window behind as well
without calling show() and causing any rendering action.

Signed-off-by: Simon Busch <morphis@gravedo.de>